# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deposits:
    get:
      operationId: getDeposits
      tags: [portfolio]
      summary: List deposit requests
      responses:
        "200":
          description: Deposit requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepositRequest"
    post:
      operationId: createDeposit
      tags: [portfolio]
      summary: Create a deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositRequest"
  /withdrawals:
    get:
      operationId: getWithdrawals
      tags: [portfolio]
      summary: List withdrawal requests
      responses:
        "200":
          description: Withdrawal requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WithdrawalRequest"
    post:
      operationId: createWithdrawal
      tags: [portfolio]
      summary: Create a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequest"
...
  /support/tickets:
    get:
      operationId: getTickets
      tags: [support]
      summary: List support tickets
      responses:
        "200":
          description: Support tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createTicket
      tags: [support]
      summary: Create a support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Support ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /support/tickets/{id}/reply:
    post:
      operationId: createTicketReply
      tags: [support]
      summary: Add a reply to a support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketReplyInput"
      responses:
        "201":
          description: Updated support ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
...
  /admin/deposits/{id}/address:
    post:
      operationId: assignDepositAddress
      tags: [admin]
      summary: Assign a manually verified TRC20 deposit address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositAddressInput"
      responses:
        "200":
          description: Updated deposit request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositRequest"
...
  /admin/withdrawals/{id}/status:
    post:
      operationId: updateWithdrawalStatus
      tags: [admin]
      summary: Update a withdrawal request status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdateInput"
      responses:
        "200":
          description: Updated withdrawal request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequest"
...
  /admin/tickets/{id}/status:
    post:
      operationId: updateTicketStatus
      tags: [admin]
      summary: Update a support ticket status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketStatusUpdateInput"
      responses:
        "200":
          description: Updated support ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /kyc:
    get:
      operationId: getKyc
      tags: [portfolio]
      summary: Get the current user's KYC submission
      responses:
        "200":
          description: KYC submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycSubmission"
    post:
      operationId: submitKyc
      tags: [portfolio]
      summary: Submit identity information for KYC review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycInput"
      responses:
        "201":
          description: KYC submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycSubmission"
...
  /admin/kyc/{id}/status:
    post:
      operationId: updateKycStatus
      tags: [admin]
      summary: Approve or reject a KYC submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycStatusUpdateInput"
      responses:
        "200":
          description: Updated KYC submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycSubmission"
  /copy-trading:
    get:
      operationId: getCopyTrading
      tags: [portfolio]
      summary: Get demo copy-trading trader and allocation
      responses:
        "200":
          description: Copy-trading demo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopyTrading"
    post:
      operationId: createCopyAllocation
      tags: [portfolio]
      summary: Set a simulated copy-trading allocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopyAllocationInput"
      responses:
        "201":
          description: Simulated allocation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopyTrading"
components:
  schemas:
...
    DepositRequest:
      type: object
      required: [id, amount, status, network, address, createdAt]
      properties:
        id: { type: number }
        amount: { type: number }
        status: { type: string, enum: [awaiting_address, awaiting_transfer, pending_review, completed, rejected] }
        network: { type: string, enum: [TRC20] }
        address: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    DepositInput:
      type: object
      required: [amount]
      properties:
        amount: { type: number, minimum: 1 }
    DepositAddressInput:
      type: object
      required: [address]
      properties:
        address: { type: string, minLength: 20 }
    WithdrawalRequest:
      type: object
      required: [id, amount, address, status, network, createdAt]
      properties:
        id: { type: number }
        amount: { type: number }
        address: { type: string }
        status: { type: string, enum: [pending_review, processing, completed, rejected] }
        network: { type: string, enum: [TRC20] }
        createdAt: { type: string, format: date-time }
    WithdrawalInput:
      type: object
      required: [amount, address]
      properties:
        amount: { type: number, minimum: 1 }
        address: { type: string, minLength: 20 }
    StatusUpdateInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending_review, processing, completed, rejected] }
...
    KycSubmission:
      type: object
      required: [id, fullName, dateOfBirth, country, documentType, documentNumber, documentReference, status, submittedAt]
      properties:
        id: { type: number }
        fullName: { type: string }
        dateOfBirth: { type: string, format: date }
        country: { type: string }
        documentType: { type: string, enum: [passport, national_id, drivers_license] }
        documentNumber: { type: string }
        documentReference: { type: string }
        status: { type: string, enum: [pending, approved, rejected] }
        rejectionReason: { type: ["string", "null"] }
        submittedAt: { type: string, format: date-time }
    KycInput:
      type: object
      required: [fullName, dateOfBirth, country, documentType, documentNumber, documentReference]
      properties:
        fullName: { type: string, minLength: 2 }
        dateOfBirth: { type: string, format: date }
        country: { type: string, minLength: 2 }
        documentType: { type: string, enum: [passport, national_id, drivers_license] }
        documentNumber: { type: string, minLength: 4 }
        documentReference: { type: string, minLength: 1 }
    KycStatusUpdateInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, approved, rejected] }
        rejectionReason: { type: ["string", "null"] }
    CopyTrading:
      type: object
      required: [trader, allocation, isSimulated]
      properties:
        trader:
          type: object
          required: [name, title, avatar, dailyReturnMin, dailyReturnMax, riskLevel, followers, strategy, performance]
          properties:
            name: { type: string }
            title: { type: string }
            avatar: { type: string }
            dailyReturnMin: { type: number }
            dailyReturnMax: { type: number }
            riskLevel: { type: string, enum: [low, medium, high] }
            followers: { type: number }
            strategy: { type: string }
            performance:
              type: array
              items: { type: number }
        allocation: { type: number }
        isSimulated: { type: boolean }
    CopyAllocationInput:
      type: object
      required: [amount]
      properties:
        amount: { type: number, minimum: 0 }
...
    SupportTicket:
      type: object
      required: [id, subject, category, message, status, createdAt, replies]
      properties:
        id: { type: number }
        subject: { type: string }
        category: { type: string }
        message: { type: string }
        status: { type: string, enum: [open, in_progress, closed] }
        createdAt: { type: string, format: date-time }
        replies:
          type: array
          items:
            type: object
            required: [id, author, message, createdAt]
            properties:
              id: { type: number }
              author: { type: string }
              message: { type: string }
              createdAt: { type: string, format: date-time }
    TicketInput:
      type: object
      required: [subject, category, message]
      properties:
        subject: { type: string, minLength: 3 }
        category: { type: string }
        message: { type: string, minLength: 10 }
    TicketReplyInput:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 2 }
    TicketStatusUpdateInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [open, in_progress, closed] }
...