# lib/api-zod/src/generated/types/copyTrading.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CopyTradingTrader } from './copyTradingTrader';

export interface CopyTrading {
  trader: CopyTradingTrader;
  allocation: number;
  isSimulated: boolean;
}