# lib/api-zod/src/generated/types/kycStatusUpdateInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KycStatusUpdateInputStatus } from './kycStatusUpdateInputStatus';

export interface KycStatusUpdateInput {
  status: KycStatusUpdateInputStatus;
  /** @nullable */
  rejectionReason?: string | null;
}