# lib/api-zod/src/generated/types/kycSubmission.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KycSubmissionDocumentType } from './kycSubmissionDocumentType';
import type { KycSubmissionStatus } from './kycSubmissionStatus';

export interface KycSubmission {
  id: number;
  fullName: string;
  dateOfBirth: Date;
  country: string;
  documentType: KycSubmissionDocumentType;
  documentNumber: string;
  documentReference: string;
  status: KycSubmissionStatus;
  /** @nullable */
  rejectionReason?: string | null;
  submittedAt: Date;
}