# lib/api-zod/src/generated/types/ticketInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TicketInput {
  /** @minLength 3 */
  subject: string;
  category: string;
  /** @minLength 10 */
  message: string;
}