# lib/api-zod/src/generated/types/withdrawalRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WithdrawalRequestNetwork } from './withdrawalRequestNetwork';
import type { WithdrawalRequestStatus } from './withdrawalRequestStatus';

export interface WithdrawalRequest {
  id: number;
  amount: number;
  address: string;
  status: WithdrawalRequestStatus;
  network: WithdrawalRequestNetwork;
  createdAt: Date;
}